grouping interfaces
Added the possibility to group interfaces the go to the same destination node. After the problem is solved, the mapping is made on the original physical network. This adds the possibility to load balance the traffic of a virtual node across multiple network interfaces (e.g., 50% on eth0 and 50% on eth1).